requirements.txt
langchain==0.1.0
python-dotenv==1.0.0
openai==1.10.0
gradio==4.12.0